+++ spec/shared/rspec_only_cop_behavior.rb
@@ -63,6 +63,6 @@ RSpec.shared_examples 'an rspec only cop', rspec_cop: true do
   end
 
   it 'deems spec/feature/thing_spec.rb to be a relevant file' do
-    expect(cop.relevant_file?('spec/feature/thing_spec.rb')).to be(true)
+    expect(cop.relevant_file?('spec/feature/thing_spec.rb')).to be_truthy
   end
 end